back-end/pyproject.toml
[project]
name = "project3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "marshmallow>=4.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "gunicorn>=21.2",
    "requests>=2.31.0"
]
